dev: add new cards controllers

+++ controllers/cards.js
@@ -2,13 +2,9 @@ const Card = require("../models/card");
 
 module.exports.getCards = (req, res) => {
   Card.find({})
-    .then((card) =>
-      res.send({
-        name: card.name,
-        avatar: card.avatar,
-        likes: card.likes,
-        owner: req.user._id,
-      })
+    .populate(["owner", "likes"])
+    .then((Cards) =>
+      res.status(200).send(Cards.sort((a, b) => b.createdAt - a.createdAt))
     )
     .catch(() =>
       res.status(500).send({
@@ -19,29 +15,31 @@ module.exports.getCards = (req, res) => {
 
 module.exports.createCard = (req, res) => {
   const { name, link } = req.body;
-  Card.create({ name, link })
-    .then((card) =>
-      res.send({
-        name: card.name,
-        link: card.link,
-        owner: req.user._id,
-      })
-    )
-    .catch(() =>
-      res.status(400).send({
-        message: "Переданы некорректные данные.",
-      })
-    );
+  Card.create({ name, link, owner: req.user._id }).then((card) => {
+    Card.findById(card._id)
+      .populate(["owner"])
+      .then((newCard) => res.status(200).send(newCard))
+      .catch(() =>
+        res.status(400).send({
+          message: "Переданы некорректные данные.",
+        })
+      );
+  });
 };
 
 module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndDelete(req.user._id)
-    .then(() => res.status(200).send({ message: "Карточка удалена успешно!" }))
-    .catch(() =>
-      res.status(500).send({
-        message: "Произошла ошибка на сервере. Сервер не отвечает.",
-      })
-    );
+  Card.findById(req.params.id).then((card) => {
+    if (card.owner.toString() === req.user._id.toString()) {
+      Card.findByIdAndDelete(req.user._id)
+        .populate(["owner", "likes"])
+        .then((deletedCard) => res.status(200).send(deletedCard))
+        .catch(() =>
+          res.status(500).send({
+            message: "Произошла ошибка на сервере. Сервер не отвечает.",
+          })
+        );
+    }
+  });
 };
 
 module.exports.likeCard = (req, res) =>
@@ -49,19 +47,23 @@ module.exports.likeCard = (req, res) =>
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
     { new: true }
-  ).catch(() =>
-    res.status(500).send({
-      message: "Произошла ошибка на сервере. Сервер не отвечает.",
-    })
-  );
+  )
+    .then((card) => res.status(200).send(card))
+    .catch(() =>
+      res.status(500).send({
+        message: "Произошла ошибка на сервере. Сервер не отвечает.",
+      })
+    );
 
 module.exports.dislikeCard = (req, res) =>
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
     { new: true }
-  ).catch(() =>
-    res.status(500).send({
-      message: "Произошла ошибка на сервере. Сервер не отвечает.",
-    })
-  );
+  )
+    .then((card) => res.status(200).send(card))
+    .catch(() =>
+      res.status(500).send({
+        message: "Произошла ошибка на сервере. Сервер не отвечает.",
+      })
+    );
